+++ telegrambots-extensions/src/main/java/org/telegram/telegrambots/bots/commandbot/TelegramLongPollingCommandBot.java
@@ -1,12 +1,15 @@
-package org.telegram.telegrambots.bots;
+package org.telegram.telegrambots.bots.commandbot;
 
 
 import org.telegram.telegrambots.ApiContext;
 import org.telegram.telegrambots.api.objects.Message;
 import org.telegram.telegrambots.api.objects.Update;
-import org.telegram.telegrambots.bots.commands.BotCommand;
-import org.telegram.telegrambots.bots.commands.CommandRegistry;
-import org.telegram.telegrambots.bots.commands.ICommandRegistry;
+import org.telegram.telegrambots.bots.AbsSender;
+import org.telegram.telegrambots.bots.DefaultBotOptions;
+import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.bots.commandbot.commands.BotCommand;
+import org.telegram.telegrambots.bots.commandbot.commands.CommandRegistry;
+import org.telegram.telegrambots.bots.commandbot.commands.ICommandRegistry;
 
 import java.util.Collection;
 import java.util.Map;